Amemd hash calculation.

+++ common.cxx
@@ -921,7 +921,7 @@ l= (int)(last-first);
 /*49:*/
 #line 656 "common.w"
 
-h= (eight_bits)*i;
+h= (int)((eight_bits)*i);
 while(++i<last)h= (h+h+(int)((eight_bits)*i))%hash_size;
